Registration and Login Systems
Successfully completed a registration and login system.

The registration system hashses sensative user credentials for storage in database. Incorporated security questions (an array stored in an outside file for access by other files) to later impliment a forgot-password and password changing system. Registration may be turned off by a True/False variable in the website configuration file.

The login system checks passwords against the stored, hashed passwords to determine if access should be granted. Logs both successful and failed sign-in attempts by means of a database table. For every success, the user is logged-in and redirect to their dashboard (dashboard.php; to be created).

This commit moves sample login and registration html pages out of the main directory and into /samples/ for reference until there are ready to be deleted.