Created Login Session
Created session_check.php to check if user is logged in.  If the user is not logged in, destroys any possible session data and redirects them to the login page.

Created logout.php to destroy the user's session. Redirects them to the login page aftewards.

Minor change to login.php to redirect the user to dashboard if they are already logged in. If they are not already logged in, a session is created after a successful login.